Add a user model and sessions#create
Allows a service to pass a temporary token from github to sessions#create which
will use the token to authenticate with github and find or create a user using
the name returned from github. This user is then returned on success.

Users will generate a random, unique token on create.